requirements.txt
Flask
pymongo
python-dotenv
cloudinary
Pillow
google-generativeai
PyJWT
Flask-Cors
werkzeug
cryptography